LPD-56159 Move method to base class
