+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/AllSectionFragmentRenderer.java
@@ -12,6 +12,7 @@ import com.liferay.object.service.ObjectDefinitionService;
 import com.liferay.object.service.ObjectDefinitionSettingLocalService;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.util.Portal;
 import com.liferay.site.cms.site.initializer.internal.display.context.AllSectionDisplayContext;
 
 import java.io.IOException;
@@ -58,7 +59,7 @@ public class AllSectionFragmentRenderer extends BaseSectionFragmentRenderer {
 				new AllSectionDisplayContext(
 					_depotEntryLocalService, _groupLocalService,
 					httpServletRequest, _language, _objectDefinitionService,
-					_objectDefinitionSettingLocalService));
+					_objectDefinitionSettingLocalService, _portal));
 
 			requestDispatcher.include(httpServletRequest, httpServletResponse);
 		}
@@ -82,6 +83,9 @@ public class AllSectionFragmentRenderer extends BaseSectionFragmentRenderer {
 	@Reference
 	private ObjectDefinitionSettingLocalService
 		_objectDefinitionSettingLocalService;
+
+	@Reference
+	private Portal _portal;
 
 	@Reference(
 		target = "(osgi.web.symbolicname=com.liferay.site.cms.site.initializer)"
